fix: use the workflows for Terraform modules when updating (#6)
# Description

The update script used the Docker workflows by accident. Changed to
`terraform_module` as this repository hosts the workflows for Terraform
modules.

Added the `release-type` parameter. Set to `manual` so that releases are
not created automatically but a PR instead.

# Verification

Script locally tested. Updated the workflows for the bastion host.